+++ src/django/api/matching.py
@@ -10,18 +10,17 @@ from datetime import datetime
 from django.conf import settings
 from django.contrib.postgres.search import TrigramSimilarity
 from django.db import transaction
-from django.db.models import Q, Max
+from django.db.models import Q
+from django.db.models.signals import post_save
 from django.utils import timezone
 
 from api.models import (Facility,
                         FacilityList,
                         FacilityListItem,
                         FacilityMatch,
-                        HistoricalFacility,
-                        HistoricalFacilityMatch,
                         TrainedModel)
 from api.helpers import clean
-from api.gazetteer import OgrGazetteer, OgrStaticGazetteer
+from api.gazetteer import OgrGazetteer, OgrStaticGazetteer, ModelOutOfDate
 
 logger = logging.getLogger(__name__)
 
@@ -353,9 +352,9 @@ def match_items(messy,
     if len(messy.keys()) > 0:
         no_geocoded_items = False
         try:
-            gazetteer = GazetteerCache.get_latest()
-            results = gazetteer.search(messy, threshold=gazetteer_threshold,
-                                       n_matches=None, generator=False)
+            results = GazetteerCache.search(
+                messy, threshold=gazetteer_threshold, n_matches=None,
+                generator=False)
             no_gazetteer_matches = False
         except NoCanonicalRecordsError:
             results = []
@@ -538,234 +537,45 @@ class GazetteerCache:
     Note that the first time `get_latest` is called it will be slow, as it
     needs to train a model and index it with all the `Facility` items.
     """
-    _gazetter = None
-    _facility_version = None
-    _match_version = None
+    _gazetteer = None
 
     @classmethod
-    def _rebuild_gazetteer(cls):
-        logger.info('Rebuilding gazetteer')
-        with transaction.atomic():
-            db_facility_version = HistoricalFacility.objects.aggregate(
-                max_id=Max('history_id')).get('max_id')
-            db_match_version = HistoricalFacilityMatch.objects.aggregate(
-                max_id=Max('history_id')).get('max_id')
-            # We expect `get_canonical_items` to return a list rather than a
-            # QuerySet so that we can close the transaction as quickly as
-            # possible
-            canonical = get_canonical_items()
-            if len(canonical.keys()) == 0:
-                raise NoCanonicalRecordsError()
-            # We expect `get_messy_items_for_training` to return a list rather
-            # than a QuerySet so that we can close the transaction as quickly
-            # as possible
-            messy = get_messy_items_for_training()
-
-        cls._gazetter = train_gazetteer(messy, canonical, should_index=True)
-        cls._facility_version = db_facility_version
-        cls._match_version = db_match_version
-        return cls._gazetter
+    def load_gazetteer_if_none(cls):
+        if cls._gazetteer is None:
+            cls._gazetteer = load_gazetteer()
+        return cls._gazetteer
 
     @classmethod
-    @transaction.atomic
-    def _get_new_facility_history(cls):
-        facility_changes = []
-        latest_facility_dedupe_records = {}
-
-        db_facility_version = HistoricalFacility.objects.aggregate(
-            max_id=Max('history_id')).get('max_id')
-
-        if db_facility_version != cls._facility_version:
-            if cls._facility_version is None:
-                last_facility_version_id = 0
-            else:
-                last_facility_version_id = cls._facility_version
-            # We call `list` so that we can get all the data and exit
-            # the transaction as soon as possible
-            facility_changes = list(
-                HistoricalFacility
-                .objects
-                .filter(history_id__gt=last_facility_version_id)
-                .order_by('history_id')
-                .extra(select={'country': 'country_code'})
-                .values('id', 'country', 'name', 'address',
-                        'history_type', 'history_id'))
-
-            changed_facility_ids_qs = HistoricalFacility \
-                .objects \
-                .filter(history_id__gt=last_facility_version_id) \
-                .values_list('id', flat=True)
-
-            # We use an dictionary comprehension so that we can load
-            # all the data and exit the transaction as soon as possible
-            latest_facility_dedupe_records = {
-                f['id']: facility_values_to_dedupe_record(f) for f in
-                Facility
-                .objects
-                .filter(id__in=changed_facility_ids_qs)
-                .extra(select={'country': 'country_code'})
-                .values('id', 'country', 'name', 'address')
-            }
-
-        return facility_changes, latest_facility_dedupe_records
+    def load_latest_gazetteer(cls):
+        cls._gazetteer = load_gazetteer()
+        return cls._gazetteer
 
     @classmethod
-    @transaction.atomic
-    def _get_new_match_history(cls):
-        match_changes = []
-        latest_match_records = {}
-        latest_matched_facility_dedupe_records = {}
-
-        db_match_version = HistoricalFacility.objects.aggregate(
-            max_id=Max('history_id')).get('max_id')
-
-        if db_match_version != cls._match_version:
-            if cls._match_version is None:
-                last_match_version_id = 0
-            else:
-                last_match_version_id = cls._match_version
-
-            # We call `list` so that we can get all the data and exit
-            # the transaction as soon as possible
-            match_changes = list(
-                HistoricalFacilityMatch
-                .objects
-                .filter(history_id__gt=last_match_version_id)
-                .order_by('history_id')
-                .values('id', 'facility', 'history_type', 'history_id'))
-
-            # We use an dictionary comprehension so that we can load
-            # all the data and exit the transaction as soon as possible
-            latest_match_records = {
-                m['id']: {
-                    'facility': m['facility'],
-                    'status': m['status'],
-                    'is_active': m['is_active'],
-                } for m in
-                FacilityMatch
-                .objects
-                .filter(id__in=(
-                    HistoricalFacilityMatch
-                    .objects
-                    .filter(history_id__gt=last_match_version_id)
-                    .values_list('id', flat=True)
-                ))
-                .values('id', 'facility', 'status', 'is_active')}
-
-            # We use an dictionary comprehension so that we can load
-            # all the data and exit the transaction as soon as possible
-            latest_matched_facility_dedupe_records = {
-                f['id']: facility_values_to_dedupe_record(f) for f in
-                Facility
-                .objects
-                .filter(id__in=(
-                    HistoricalFacilityMatch
-                    .objects
-                    .filter(history_id__gt=last_match_version_id)
-                    .values_list('facility', flat=True)
-                ))
-                .extra(select={'country': 'country_code'})
-                .values('id', 'country', 'name', 'address')
-            }
-
-        return (match_changes, latest_match_records,
-                latest_matched_facility_dedupe_records)
-
-    @classmethod
-    def get_latest(cls):
+    def index(cls, data):
+        gazetteer = cls.load_gazetteer_if_none()
         try:
-            if cls._gazetter is None:
-                return load_gazetteer()
+            return gazetteer.index(data)
+        except ModelOutOfDate:
+            gazetteer = cls.load_latest_gazetteer()
+            return gazetteer.index(data)
 
-            facility_changes, latest_facility_dedupe_records = \
-                cls._get_new_facility_history()
+    @classmethod
+    def unindex(cls, data):
+        gazetteer = cls.load_gazetteer_if_none()
+        try:
+            return gazetteer.unindex(data)
+        except ModelOutOfDate:
+            gazetteer = cls.load_latest_gazetteer()
+            return gazetteer.unindex(data)
 
-            for item in facility_changes:
-                # We were previously calling `cls._gazetter.unindex` to
-                # remove records with a `history_type` of `-` but it was
-                # raising exceptions for which we could not determine the
-                # root cause. We have opted to ignore them and filter out
-                # no longer existing records from the match results.
-                if item['history_type'] != '-':
-                    # The history record has old field values, so we use the
-                    # updated version that we fetched. If we don't have a
-                    # record for the ID, it means that the facility has been
-                    # deleted. We don't need to index a deleted facility.
-                    if item['id'] in latest_facility_dedupe_records:
-                        record = latest_facility_dedupe_records[item['id']]
-                        logger.debug(
-                            'Indexing facility {}'.format(str(record)))
-                        cls._gazetter.index(record)
-                cls._facility_version = item['history_id']
-
-            (match_changes, latest_match_records,
-             latest_matched_facility_dedupe_records) = \
-                cls._get_new_match_history()
-
-            def dedupe_record_for_match_item(item):
-                facility_id = item['facility']
-                key = match_detail_to_extended_facility_id(
-                    facility_id, item['id'])
-                """
-                The latest_matched_facility_dedupe_records dictionary looks
-                like this:
-
-                {
-                    facility_id: {
-                        facility_id: {
-                            field1: value1,
-                            field2, value2
-                        }
-                    }
-                }
-
-                We want to get the inner object and change the key from a real
-                facility ID to a "synthetic" facility ID which we use to index
-                confirmed matches.
-                """
-                value = (
-                    latest_matched_facility_dedupe_records
-                    [facility_id][facility_id]
-                )
-                return {key: value}
-
-            for item in match_changes:
-                match = (latest_match_records[item['id']]
-                         if item['id'] in latest_match_records
-                         else None)
-                has_facility = (
-                    item['facility'] in latest_matched_facility_dedupe_records)
-                is_confirmed_match_with_facility = (
-                    match
-                    and match['status'] == FacilityMatch.CONFIRMED
-                    and has_facility)
-                if is_confirmed_match_with_facility:
-                    # We were previously calling `cls._gazetter.unindex` to
-                    # remove records with a `history_type` of `-` but it was
-                    # raising exceptions for which we could not determine the
-                    # root cause. We have opted to ignore them and filter out
-                    # no longer existing records from the match results.
-                    if item['history_type'] != '-':
-                        # The history record has old field values, so we us the
-                        # updated version that we fetched. If we don't have a
-                        # record for the ID, it means that the facility has
-                        # been deleted. We don't need to index a deleted
-                        # facility.
-                        if match and match['is_active']:
-                            record = dedupe_record_for_match_item(item)
-                            logger.debug(
-                                'Indexing match {}'.format(str(record)))
-                            cls._gazetter.index(record)
-                    cls._match_version = item['history_id']
-
-        except Exception:
-            extra_info = {
-                'last_successful_facility_version': cls._facility_version,
-                'last_successful_match_version': cls._match_version}
-            _try_reporting_error_to_rollbar(extra_info)
-            raise
-
-        return cls._gazetter
+    @classmethod
+    def search(cls, messy, threshold, n_matches, generator):
+        gazetteer = cls.load_gazetteer_if_none()
+        try:
+            return gazetteer.search(messy, threshold, n_matches, generator)
+        except ModelOutOfDate:
+            gazetteer = cls.load_latest_gazetteer()
+            return gazetteer.search(messy, threshold, n_matches, generator)
 
 
 def get_model_data():
@@ -782,3 +592,41 @@ def get_model_data():
         messy = get_messy_items_for_training()
 
     return messy, canonical
+
+
+def should_index_match_with_dedupe(match):
+    if match.is_active:
+        if match.status == FacilityMatch.AUTOMATIC:
+            return match.facility.created_from == match.facility_list_item
+        if match.status == FacilityMatch.CONFIRMED:
+            return True
+    return False
+
+
+def should_unindex_match_from_dedupe(match):
+    if not match.is_active:
+        return match.facility.created_from != match.facility_list_item
+    return False
+
+
+def dedupe_record_for_match(match):
+    record_id = (
+        match.facility.id
+        if match.facility.created_from == match.facility_list_item
+        else match_to_extended_facility_id(match))
+    return {record_id: {
+        'country': clean(match.facility_list_item.country_code),
+        'name': clean(match.facility_list_item.name),
+        'address': clean(match.facility_list_item.address),
+    }}
+
+
+def facilitymatch_post_save(sender, **kwargs):
+    instance = kwargs.get('instance')
+    if should_index_match_with_dedupe(instance):
+        GazetteerCache.index(dedupe_record_for_match(instance))
+    if should_unindex_match_from_dedupe(instance):
+        GazetteerCache.unindex(dedupe_record_for_match(instance))
+
+
+post_save.connect(facilitymatch_post_save, sender=FacilityMatch)